changes to ChessGame.ai method

+++ src/main/java/chess/player/ai/uci/engine/UCIEngine.java
@@ -105,10 +105,12 @@ abstract class UCIEngine {
         try {
             sendCommand("quit");
         } finally {
-            process.destroyForcibly();
+            process.destroy();
+            if (process.isAlive()) {
+                process.destroyForcibly();
+            }
             input.close();
             output.close();
-            System.gc();
         }
     }
